chore(contentful): tests moved to folders
